It renders!

+++ src/index.js
@@ -15,7 +15,10 @@ async function main() {
     [0, 1],
     [1, 0],
   ];
-  const grid = Uint8Array.from(jsGrid.flat());
+  // kinda dumb we gotta use a 32bit array to store a boolean value. WGSL supports array<bool>,
+  // but there isn't an equivalent in JS that we can use to transfer it. Maybe we'll experiment with
+  // binpacking at some point, but not now.
+  const grid = Uint32Array.from(jsGrid.flat());
   const width = jsGrid[0].length;
   const height = jsGrid.length;
